add 0.5 estimates in employer review

+++ backend/src/main/java/ru/hh/school/employerreview/review/ReviewResource.java
@@ -65,7 +65,7 @@ public class ReviewResource {
         reviewDto.getText());
     reviewDao.save(review);
 
-    ratingDao.addNewEstimate(employer.getId(), reviewDto.getRating().intValue());
+    ratingDao.addNewEstimate(employer.getId(), reviewDto.getRating());
 
     return Response.status(200).entity(new ResponseBodyReviewId(review.getId())).build();
   }
